Move executeCommand out of Cline and add telemetry for shell integration errors
