+++ .changeset/four-trainers-move.md
@@ -1,0 +1,5 @@
+---
+"roo-cline": patch
+---
+
+Move executeCommand out of Cline and add telemetry for shell integration errors
