+++ src/core/__tests__/read-file-xml.test.ts
@@ -121,7 +121,8 @@ describe("read_file tool XML output structure", () => {
 		mockCline.getFileContextTracker = jest.fn().mockReturnValue({
 			trackFileContext: jest.fn().mockResolvedValue(undefined),
 		})
-		mockCline.recordToolUsage = jest.fn().mockReturnValue({} as ToolUsage)
+		mockCline.recordToolUsage = jest.fn().mockReturnValue(undefined)
+		mockCline.recordToolError = jest.fn().mockReturnValue(undefined)
 
 		// Reset tool result
 		toolResult = undefined
