+++ src/core/prompts/tools/index.ts
@@ -1,6 +1,5 @@
 import { ToolName } from "../../../schemas"
-import { TOOL_GROUPS, ALWAYS_AVAILABLE_TOOLS } from "../../../shared/tools"
-import { DiffStrategy } from "../../diff/DiffStrategy"
+import { TOOL_GROUPS, ALWAYS_AVAILABLE_TOOLS, DiffStrategy } from "../../../shared/tools"
 import { McpHub } from "../../../services/mcp/McpHub"
 import { Mode, ModeConfig, getModeConfig, isToolAllowedForMode, getGroupName } from "../../../shared/modes"
