+++ src/services/telemetry/PostHogClient.ts
@@ -32,6 +32,7 @@ export class PostHogClient {
 		ERRORS: {
 			SCHEMA_VALIDATION_ERROR: "Schema Validation Error",
 			DIFF_APPLICATION_ERROR: "Diff Application Error",
+			SHELL_INTEGRATION_ERROR: "Shell Integration Error",
 			CONSECUTIVE_MISTAKE_ERROR: "Consecutive Mistake Error",
 		},
 	}
